+++ src/main/java/net/enzo/plus/client/render/InfinityChestRenderItem.java
@@ -4,12 +4,16 @@ import cpw.mods.fml.relauncher.Side;
 import cpw.mods.fml.relauncher.SideOnly;
 import net.enzo.plus.common.tiles.TileInfinityChest;
 import net.enzo.plus.common.tiles.TileInfinityChestRenderer;
+import net.minecraft.client.model.ModelChest;
 import net.minecraft.item.ItemStack;
 import net.minecraftforge.client.IItemRenderer;
+import org.lwjgl.opengl.GL11;
+import org.lwjgl.opengl.GL12;
 
 @SideOnly(Side.CLIENT)
 public class InfinityChestRenderItem implements IItemRenderer {
     private static final TileInfinityChest tileEntityInfinityChest = new TileInfinityChest();
+    ModelChest model = new ModelChest();
 
     @Override
     public boolean handleRenderType(ItemStack item, ItemRenderType type)
@@ -23,12 +27,11 @@ public class InfinityChestRenderItem implements IItemRenderer {
         return true;
     }
 
+    @SideOnly(Side.CLIENT)
     @Override
     public void renderItem(ItemRenderType type, ItemStack item, Object... data)
     {
-        if (type == ItemRenderType.ENTITY)
-            TileInfinityChestRenderer.instance.renderTileEntityAt(tileEntityInfinityChest, -0.5F, -0.5F, -0.5F, 0);
-        else
-            TileInfinityChestRenderer.instance.renderTileEntityAt(tileEntityInfinityChest, 0, 0, 0, 0);
+        model = new ModelChest();
+        model.renderAll();
     }
 }