+++ src/main/java/net/enzo/plus/common/item/tools/ItemInfinityPickaxe.java
@@ -27,7 +27,7 @@ public class ItemInfinityPickaxe extends ItemPickaxe {
 
     public ItemInfinityPickaxe() {
         super(opPick);
-        setUnlocalizedName("infinity_pickaxe");
+        setUnlocalizedName("infinity_rpickaxe");
         setTextureName("plus:infinity_pickaxe");
         setCreativeTab(PlusMod.tab);
         setMaxDamage(0);